*** empty log message ***

+++ web_module/branches/new_or_2006_07_25/src/main/com/topcoder/apps/review/projecttracker/ProjectTrackerV2Bean.java
@@ -804,7 +804,7 @@ public class ProjectTrackerV2Bean implements SessionBean {
             double old_rating = 0;
 
             if (rs.next()) {
-                old_rating = rs.getDouble(1);
+                old_rating = rs.getLong(1);
             }
 
             Common.close(rs);
